load secret key file from s3
