+++ src/connector/Cargo.toml
@@ -19,7 +19,7 @@ apache-avro = { git = "https://github.com/risingwavelabs/avro", branch = "waruto
 async-trait = "0.1"
 aws-config = { workspace = true }
 aws-sdk-kinesis = { workspace = true } 
-aws-sdk-s3 = { workspace = true } 
+aws-sdk-s3 = { workspace = true }
 aws-smithy-http = { workspace = true }
 aws-types = { workspace = true }
 bincode = "1"
@@ -60,6 +60,7 @@ serde_derive = "1"
 serde_json = "1"
 serde_with = "2"
 simd-json = { git = "https://github.com/tabVersion/simd-json.git", branch = "main", features = ["key-to-lowercase"] }
+tempfile = "3"
 thiserror = "1"
 tokio = { version = "0.2", package = "madsim-tokio", features = ["rt", "rt-multi-thread", "sync", "macros", "time", "signal", "fs"] }
 tokio-retry = "0.3"
@@ -74,5 +75,4 @@ workspace-hack = { path = "../workspace-hack" }
 
 [dev-dependencies]
 rand = "0.8"
-tempfile = "3"
 wiremock = "0.5"
