+++ src/connector/src/aws_utils.rs
@@ -26,7 +26,10 @@ use aws_smithy_http::endpoint::Endpoint;
 use aws_types::credentials::SharedCredentialsProvider;
 use aws_types::region::Region;
 use http::Uri;
+use risingwave_common::error::ErrorCode::{InternalError, InvalidConfigValue, ProtocolError};
+use risingwave_common::error::{Result, RwError};
 use serde::{Deserialize, Serialize};
+use url::Url;
 
 pub const AWS_DEFAULT_CONFIG: [&str; 7] = [
     "region",
@@ -300,6 +303,45 @@ impl AwsConfigV2 {
             ),
         }
     }
+}
+
+// TODO(Tao): Probably we should never allow to use S3 URI.
+/// properties require keys: refer to [`AWS_DEFAULT_CONFIG`]
+pub async fn load_file_descriptor_from_s3(
+    location: &Url,
+    properties: &HashMap<String, String>,
+) -> Result<Vec<u8>> {
+    let bucket = location.domain().ok_or_else(|| {
+        RwError::from(InternalError(format!(
+            "Illegal Protobuf schema path {}",
+            location
+        )))
+    })?;
+    // if properties.get(PB_SCHEMA_LOCATION_S3_REGION).is_none() {
+    //     return Err(RwError::from(InvalidConfigValue {
+    //         config_entry: PB_SCHEMA_LOCATION_S3_REGION.to_string(),
+    //         config_value: "NONE".to_string(),
+    //     }));
+    // }
+    let key = location.path().replace('/', "");
+    let config = AwsConfigV2::from(properties.clone());
+    let sdk_config = config.load_config(None).await;
+    let s3_client = s3_client(&sdk_config, Some(default_conn_config()));
+    let response = s3_client
+        .get_object()
+        .bucket(bucket.to_string())
+        .key(&key)
+        .send()
+        .await
+        .map_err(|e| RwError::from(InternalError(e.to_string())))?;
+
+    let body = response.body.collect().await.map_err(|e| {
+        RwError::from(InternalError(format!(
+            "Read Protobuf schema file from s3 {}",
+            e
+        )))
+    })?;
+    Ok(body.into_bytes().to_vec())
 }
 
 #[cfg(test)]
